third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFESpotLightElement-svgdom-pointsAtZ-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'pointsAtZ' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.pointsAtZ.baseVal is 100
PASS spotLightElement.pointsAtZ.baseVal is 0
PASS successfullyParsed is true

TEST COMPLETE
